Move base alignment operations into alignment.py
